(Re)move results  (#68)
* move results to algo files

* remove solution/qualitymetrics for now

they are safely stored on another branch but are currently not used so there is no reason to have them in this release

* move the results objects to algorithms

* fix cyclic deps
