+++ .pylintdict
@@ -541,7 +541,6 @@ simonetto
 sj
 sklearn
 slsqp
-SolutionInterface
 solutionstatus
 spsa
 sqrt
